Added more test scenarios to cover different paths how the states change
This is overall an interesting story. It seems that with this kind of
implementation mutant has it easier to discover the lacking coverage.
Probably we can/should refactor our tests to better expose that it's
a state machine. We could prepare different ways to get to a certain
state then when we want to test a transition, the tests should cover all
permutations of the state preparation for each state transition.

So, without those tests, the polymorphic implementation had 18 alive
mutants. Now it has only the 2 ones, which other implementations seem to
have too (version and stream name).